backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
bcrypt>=4.2.0
pyjwt>=2.9.0
python-multipart>=0.0.9
aiosmtplib>=3.0.0
httpx>=0.27.0
apscheduler>=3.10.0
pydantic-settings>=2.4.0
python-dateutil>=2.9.0
alembic>=1.13.0
pytest>=8.3.0